Remove the slave ID argument of modbus_new_rtu()
modbus_set_slave must be used to set the slave ID of the remote
device to talk in master mode and to set the internal slave ID in
slave mode.

If you talk to several devices, you need to call modbus_set_slave
each time the following requests must be sent to another device.