configure: Correctly detect if we are cross-compiling for win32
First, it's $host_os which should be checked, not $target (remember,
'build' is where we build, host is where the program will be run, and
target is only for compiler - for which system the compiler will
generate code.

Second, correct OS pattern - e.g. on debian with i586-mingw32msvc-gcc,
host_os is mingw32msvc, so we better for *mingw32*.

Also, it's better to call AC_CANONICAL_* for normalizing variables.

Build tested with `configure --host=i586-mingw32msvc` on Debian
GNU/Linux 5.0 (Lenny).

Cc: Александр Сёмуха <sav@mns.spb.ru>
Signed-off-by: Kirill Smelkov <kirr@mns.spb.ru>
Signed-off-by: Stéphane Raimbault <stephane.raimbault@gmail.com>