Fix unit identifier not copied by the TCP server
Reported by Antti Manninen and according to Page6 in the document
"MODBUS Messaging on TCP/IP Implementation Guide V1.0b",
------------------------
Unit Identifier – This field is used for intra-system routing purpose.
It is typically used to communicate to a MODBUS+ or a MODBUS serial
line slave through a gateway between an Ethernet TCP-IP network and
a MODBUS serial line. This field is set by the MODBUS Client in the
request and must be returned with the same value in the response by
the server.
------------------------