Modbus/TCP: Implement IPv6 support.
The new function modbus_new_tcp_pi() takes an IPv6 address or a hostname
in addition to an IPv4 address as the "node" argument. If the node is
unspecified (using a NULL pointer), modbus_tcp_listen() will bind to the
"any" address.

The "service" argument of the new function is a string argument which
supports service names in addition to port numbers.

The old function is still exported in order to maintain ABI compatibility.
Legacy applications can request the old API by defining "MODBUS_LEGACY"
before including the libmodbus header files. If the define doesn't exist,
a macro aliases modbus_new_tcp() to modbus_new_tcp_pi().

Signed-off-by: Florian Forster <octo@noris.net>