New function modbus_send_raw_request
The _modbus_send_raw_request()_ function shall send a request via
the socket of the context 'ctx'. This function must be used for
debugging purposes because you have to take care to make a valid
request by hand. The function only adds to the message, the header
or CRC of the selected backend, so 'raw_req' must start and contain
at least a slave/unit identifier and a function code. This function
can be used to send request not handled by the library.