Added a new function named modbus_last_error.
This is the same function as modbus_strerror. The implementation allows wrapping of libmodbus for .Net/C#.
It avoids two problems. First there is no way to return a char * for .Net (at least I have
found no way). Second the global errno is rather difficult to use (may have a wrong value, too).
The solution is the new method that copies the modbus_strerror return value and the error into
the given parameter err_msg.

The whole code is encapsulated into #if defined(CSHARPWRAPPER).

In modbus-tcp.c the errno is explicitly set using the error code from the winsock
implementation. This may be useful for any Win32 code. But I am not sure.
Therefor I have wrapped it the same way as the other code.