Port libmodbus to .Net/C#.
libmodbus is wrapped for use in .Net using C#. The library itself remains unchanged.
The provided wrapper class encapsulates the required marshaling to access unmanaged code
from managed code. Enumerations and Constants are provided. Additionally the naming
conventions of C# are respected. The documentation of the methods is mostly copied from
the related functions in modbus.c. It is shown from Intellisense.

The focus was an the client side. Therefore the UnitTestClient.Net is the ported
unit-test-client.c. It runs successfully against the C unit-test-server on Win7, 32 bit,
and on Win 64 bit, but as 32 bit executable.