[docs] Make clear that memory must be allocated
Memory must be allocated when using the report slave id function, this is
not handled by the function. To be safe when the length of the response is
unknown you can use the constant MODBUS_RTU_MAX_ADU_LENGTH
defined in modbus-rtu.h.

Update documentation just to help people and to point to the useful constant.