Do not reply on broadcast requests (fixes #153)
According to the Modbus specification
(http://www.modbus.org/docs/Modbus_over_serial_line_V1_02.pdf, section 2.1)
a Modbus RTU master can send a broadcast to all of it's slaves. This
broadcasts can only be write requests as otherwise collisions could
occur, eg. on a RS-485 bus.

When receiving such a broadcast, the slave should process the request as
usual, but must not reply anything, neither a normal response nor an
exception reply in case of an error.

Adjust the unit test for this case, too.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>