Bumping current LD version from 6 to 7
The CURRENT LD version for libmodbus has not changed
between v3.11 and 3.14.
But the interface for libmodbus has changed in the
set_get_timeout functions. (uses sec, usec instead of struct ts)

The rule for CURRENT LD is:
If any interfaces have been added, removed,
or changed since the last update, increment current,
and set revision to 0.