Lengthened timeout on unit tests
Discovered that unit test was often failing on ARM hardware.  Tracked issue down to how long the unit-test-server was sleeping.  It seems that when the server sleeps for 5ms between sending bytes the unit-test-server does not always get bytes in time(server generally sleeps for almost 6ms with a timeout of 10ms on the client)