CodingStyle: add .clang-format and the derived changes
This commit adds a .clang-format-file which helps to automate
the code-formatting. This file is as close as it gets to the existing
style. Some things are not supported by clang-format yet
(mainly PreProcessorIndent - currently under code-review).

This commit does not necessarily needs to be merged, but it
helps to show inconsistencies.