Add support for local echo suppression
Usually, RS-485 hardware for half-duplex connections is designed to
suppress self-generated TX traffic on the RX side. This is usually
achived by controlling the Receive Enable pin of the driver with
the same signal which toggles the Transmit Enable pin (traditionally
one pin uses an inverted signal, so this is really easy to implement).

However, some boards are designed differently in that they enable RX side
unconditionally or make this configurable with e.g. a jumper.
In this case you'll see all of your outgoing traffic also on
incoming side. So to filter out this expected but unwanted traffic
we could simply flush the incoming queue. But this way, we could
drop erroneously bytes of the legitimate response from a peer, e.g.
if the peer answers very quickly.

The approach implemented here makes use of the knowledge how many bytes
are expected and thus must be dropped. Do it directly in the transmitting
function to prevent handing over the expected byte count through
modbus context.

Also add required functions to enable/disable this workaround and
provide API documentation.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>