Implement function (0x2B / 0x0E) Read Device Identification.
The function is defined in section 6.21 of the Modbus Application Protocol v1.1b.
There are many ways to read the objects. The one used here is code 01, which seems
to be the most widely supported one.

It might be necessary to make many requests in order to receive all the data. This
is all handled internally.