Have autoreconf fail if LT_INIT is not defined
If the LT_INIT macro is not available in a *.m4 file to be
found during the autoreconf run, it is impossible to build
a useful configure script.

Therefore, we blacklist the LT_INIT macro so when it appears
unexpanded in the configure script, autoreconf will report
the failure instead of silently creating a broken configure.

This is generally a good idea for any macro not shipped with
Automake or Autoconf.