Check if ai_list is null before freeing it
Passing null to freeaddrinfo is undefined behaviour. While glibc and uClibc both handle it safely,
musl does not, so a bad nodename or servname causes a segfault.

This change ensures that the ai_list is non-null before trying to free it. In musl and glibc this
condition will never be met, but there are some paths for it to be met in uClibc.